--- JjanguMall/src/main/java/com/saltlux/jjangulmall/service/AdminService.java
@@ -1,25 +1,0 @@
-package com.saltlux.jjangulmall.service;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.saltlux.jjangumall.dto.AdminDTO;
-import com.saltlux.jjangumall.repository.AdminRepository;
-
-@Service
-public class AdminService {
-
-	@Autowired
-	private AdminRepository adminRepository;
-
-	public AdminDTO loginProcess(String id, String pwd) {
-		Map<String, String> map = new HashMap<String, String>();
-		map.put("id", id);
-		map.put("pwd", pwd);
-		return adminRepository.loginProcess(map);
-	}
-	
-}
