+++ JjanguMall/src/main/java/com/saltlux/jjangumall/security/LoginInterceptor.java
@@ -11,13 +11,13 @@ import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
 
 import com.saltlux.jjangumall.dto.UserDTO;
-import com.saltlux.jjangumall.service.store.UserService;
+import com.saltlux.jjangumall.service.store.SongUserService;
 
 
 public class LoginInterceptor extends HandlerInterceptorAdapter {
 	
 	@Autowired
-	private UserService userService;
+	private SongUserService userService;
 
 	@Autowired
 	private PasswordEncoder passwordEncoder;
