+++ JjanguMall/src/main/java/com/saltlux/jjangumall/service/store/SongUserService.java
@@ -6,13 +6,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.saltlux.jjangumall.dto.UserDTO;
-import com.saltlux.jjangumall.repository.store.UserRepository;
+import com.saltlux.jjangumall.repository.store.SongUserRepository;
 
 @Service
-public class UserService {
+public class SongUserService {
 
 	@Autowired
-	private UserRepository userRepository;
+	private SongUserRepository userRepository;
